Completed AtomicBitFieldVec

+++ tests/bit_field_vec.rs
@@ -73,7 +73,9 @@ fn test_atomic_bit_field_vec() {
 
         let cp = AtomicBitFieldVec::new(bit_width, n);
         for _ in 0..10 {
-            let values = (0..n).map(|_| rng.gen_range(0..u)).collect::<Vec<_>>();
+            let values = (0..n)
+                .map(|_| rng.gen_range(0..u) as usize)
+                .collect::<Vec<_>>();
 
             let mut indices = (0..n).collect::<Vec<_>>();
             indices.shuffle(&mut rng);
